#1645 Support worker roles for Azure Adds Microsoft.PythonTools.Worker.targets file that extends Web.targets to support worker roles. Significant improvements to Web.targets to simplify deployment to Cloud Service and Web Sites. In particular, we no longer auto-generate different web.config files for Web Sites vs. Cloud Service, but for Web Sites we will include a parameters.xml file that rewrites the environment variables on deployment. Rewrites AzureSetup into a Powershell script, but keeps the old AzureSetup around for projects that have not yet added the new scripts. AzureSetup/ConfigureCloudService will auto-download WebPI if necessary, with an option to deploy the installer manually if desired. ConfigureCloudService will auto-download pip and install a requirements.txt if necessary, with an option to deploy the downloader and sources manually if desired. Adds templates for a simple Worker Role and the support files necessary for deployment. Suppresses template name generation for certain names (bin, *.manifest, *.config) in line with other project systems.
#2006 Cloud service config should target newer osfamily
Updates PythonProjectNode to return target framework based on the project file.
Updates Web.targets to default to 4.5.1

Also mitigates a rare race condition in DerivedInterpreterFactory.
Fixes assembly signing for assemblies that should not be delay-signed.

+++ Python/Product/PythonTools/pip_downloader.py
@@ -41,7 +41,10 @@ try:
     os.chdir(setuptools_temp_dir)
     print('Downloading setuptools from ' + SETUPTOOLS_SOURCE)
     sys.stdout.flush()
-    setuptools_package, _ = urlretrieve(SETUPTOOLS_SOURCE, 'setuptools.tar.gz')
+    if os.path.exists('setuptools.tar.gz'):
+        setuptools_package = 'setuptools.tar.gz'
+    else:
+        setuptools_package, _ = urlretrieve(SETUPTOOLS_SOURCE, 'setuptools.tar.gz')
 
     package = tarfile.open(setuptools_package)
     try:
@@ -63,7 +66,10 @@ try:
     os.chdir(pip_temp_dir)
     print('Downloading pip from ' + PIP_SOURCE)
     sys.stdout.flush()
-    pip_package, _ = urlretrieve(PIP_SOURCE, 'pip.tar.gz')
+    if os.path.exists('pip.tar.gz'):
+        pip_package = 'pip.tar.gz'
+    else:
+        pip_package, _ = urlretrieve(PIP_SOURCE, 'pip.tar.gz')
 
     package = tarfile.open(pip_package)
     try:
